Prevent target sighash collision

+++ src/v2/TargetSighashLib.sol
@@ -13,7 +13,8 @@ library TargetSighashLib {
         address target,
         bytes4 selector
     ) internal pure returns (TargetSighash targetSighash) {
-        targetSighash =
-            TargetSighash.wrap(uint160(target) << 32 | uint32(selector));
+        targetSighash = TargetSighash.wrap(
+            (uint256(uint160(target)) << 32) | uint32(selector)
+        );
     }
 }
